test: PreviewAction feature test. Cleanup.
