+++ tests/src/Filament/Resources/PageResource/Pages/CreatePage.php
@@ -3,11 +3,31 @@
 namespace Pboivin\FilamentPeek\Tests\Filament\Resources\PageResource\Pages;
 
 use Filament\Resources\Pages\CreateRecord;
+use Pboivin\FilamentPeek\Pages\Actions\PreviewAction;
+use Pboivin\FilamentPeek\Pages\Concerns\HasPreviewModal;
 use Pboivin\FilamentPeek\Tests\Filament\Resources\PageResource;
 
 class CreatePage extends CreateRecord
 {
-    use HasPagePreview;
+    use HasPreviewModal;
 
     protected static string $resource = PageResource::class;
+
+    protected function getActions(): array
+    {
+        return [
+            PreviewAction::make()
+                ->label('Test_Preview_Action'),
+        ];
+    }
+
+    protected function getPreviewModalView(): ?string
+    {
+        return 'preview-page';
+    }
+
+    protected function getPreviewModalDataRecordKey(): ?string
+    {
+        return 'page';
+    }
 }
